+++ app/src/main/java/es/murallaromana/MiguelVazquezpmdm/proyecto/model/entidades/Pelicula.kt
@@ -1,5 +1,6 @@
 package es.murallaromana.MiguelVazquezpmdm.proyecto.model.entidades
 
+
 import java.io.Serializable
 
 
@@ -7,7 +8,7 @@ class Pelicula (
     var id:String?,
     var title :String,
     var rating: Number,
-    var runtimeMinutes: Int,
+    var runtimeMinutes: Double,
     var genre: String,
     var directorFullname : String,
     var imageUrl : String,
